Fix skip behaviour and recursive copying
Previously setting bundling to `undefined` would make the AssetStaging in CDK
copy the rootDir into the cdk.out folder. If the cdk.out folder was part of the
rootDir then this caused recursive copying and an eventual error.

Calling `cdk list` would trigger this as it sets `bundlingRequired` to `false`
which sets `skip` to true.

The code now handles `skip` being true and doesn't try to find Python versions
or do anything that would require the actual asset contents to be there.

Tests are updated to align more with how the code is called when ran by CDK.