Add git based bundling strategy
If we're not using a package in a workspace app then use the last commit hash
for the entire rootDir and anything in git diff to generate the custom hash for
CDK bundling.

If we are using a package in a workspace app then get the list of dependencies
and check and local ones for changes. This uses the hash of the last commit and
any git diff in that part of the tree for each dependency and the package
itself.